Updated with suggestions from code reviewers, and tobac v1.3.
Updated to tobac v1.3.2, and made suggested changes. Merge_split.py now ingests a pandas dataframe, ancillary functions have been moved to utils.py.
