+++ tobac/merge_split.py
@@ -8,225 +8,16 @@
     
 """
 
-try:
-    import geopy
-except ImportError:
-    geopy = None
 
-if geopy:
-    from geopy.distance import geodesic
-else:
-    print("You could be merge/splitting in lat/lons if you had geopy.")
-
-try:
-    import networkx
-except ImportError:
-    networkx = None
-
-if networkx:
-    from networkx import *
-else:
-    print("Cannot Merge/Split. Please install networkx.")
-# from geopy.distance import geodesic
-# from networkx import *
-import numpy as np
-from pandas.core.common import flatten
-import xarray as xr
-
-
-def compress_all(nc_grids, min_dims=2):
-    """
-    The purpose of this subroutine is to compress the netcdf variables as they are saved
-    this does not change the data, but sets netcdf encoding parameters.
-    We allocate a minimum number of dimensions as variables with dimensions
-    under the minimum value do not benefit from tangibly from this encoding.
-
-    Parameters
-    ----------
-    nc_grids : xarray.core.dataset.Dataset
-        Xarray dataset that is intended to be exported as netcdf
-
-    min_dims : integer
-        The minimum number of dimesnions, in integer value, a variable must have in order
-        set the netcdf compression encoding.
-
-    Returns
-    -------
-    nc_grids : xarray.core.dataset.Dataset
-        Xarray dataset with netcdf compression encoding for variables with two (2) or more dimensions
-
-    """
-
-    for var in nc_grids:
-        if len(nc_grids[var].dims) >= min_dims:
-            # print("Compressing ", var)
-            nc_grids[var].encoding["zlib"] = True
-            nc_grids[var].encoding["complevel"] = 4
-            nc_grids[var].encoding["contiguous"] = False
-    return nc_grids
-
-
-def standardize_track_dataset(TrackedFeatures, Mask, Projection=None):
-    """
-    Combine a feature mask with the feature data table into a common dataset.
-
-    returned by tobac.themes.tobac_v1.segmentation
-    with the TrackedFeatures dataset returned by tobac.themes.tobac_v1.linking_trackpy.
-
-    Also rename the variables to be more desciptive and comply with cf-tree.
-
-    Convert the default cell parent ID  to an integer table.
-
-    Add a cell dimension to reflect
-
-    Projection is an xarray DataArray
-
-    TODO: Add metadata attributes
-
-    Parameters
-    ----------
-    TrackedFeatures : xarray.core.dataset.Dataset
-        xarray dataset of tobac Track information, the xarray dataset returned by tobac.tracking.linking_trackpy
-
-    Mask: xarray.core.dataset.Dataset
-        xarray dataset of tobac segmentation mask information, the xarray dataset returned
-        by tobac.segmentation.segmentation
-
-
-    Projection : xarray.core.dataarray.DataArray, default = None
-        array.DataArray of the original input dataset (gridded nexrad data for example).
-        If using gridded nexrad data, this can be input as: data['ProjectionCoordinateSystem']
-        An example of the type of information in the dataarray includes the following attributes:
-        latitude_of_projection_origin :29.471900939941406
-        longitude_of_projection_origin :-95.0787353515625
-        _CoordinateTransformType :Projection
-        _CoordinateAxes :x y z time
-        _CoordinateAxesTypes :GeoX GeoY Height Time
-        grid_mapping_name :azimuthal_equidistant
-        semi_major_axis :6370997.0
-        inverse_flattening :298.25
-        longitude_of_prime_meridian :0.0
-        false_easting :0.0
-        false_northing :0.0
-
-    Returns
-    -------
-
-    ds : xarray.core.dataset.Dataset
-        xarray dataset of merged Track and Segmentation Mask datasets with renamed variables.
-
-    """
-    feature_standard_names = {
-        # new variable name, and long description for the NetCDF attribute
-        "frame": (
-            "feature_time_index",
-            "positional index of the feature along the time dimension of the mask, from 0 to N-1",
-        ),
-        "hdim_1": (
-            "feature_hdim1_coordinate",
-            "position of the feature along the first horizontal dimension in grid point space; a north-south coordinate for dim order (time, y, x)."
-            "The numbering is consistent with positional indexing of the coordinate, but can be"
-            "fractional, to account for a centroid not aligned to the grid.",
-        ),
-        "hdim_2": (
-            "feature_hdim2_coordinate",
-            "position of the feature along the second horizontal dimension in grid point space; an east-west coordinate for dim order (time, y, x)"
-            "The numbering is consistent with positional indexing of the coordinate, but can be"
-            "fractional, to account for a centroid not aligned to the grid.",
-        ),
-        "idx": ("feature_id_this_frame",),
-        "num": (
-            "feature_grid_cell_count",
-            "Number of grid points that are within the threshold of this feature",
-        ),
-        "threshold_value": (
-            "feature_threshold_max",
-            "Feature number within that frame; starts at 1, increments by 1 to the number of features for each frame, and resets to 1 when the frame increments",
-        ),
-        "feature": (
-            "feature_id",
-            "Unique number of the feature; starts from 1 and increments by 1 to the number of features",
-        ),
-        "time": (
-            "feature_time",
-            "time of the feature, consistent with feature_time_index",
-        ),
-        "timestr": (
-            "feature_time_str",
-            "String representation of the feature time, YYYY-MM-DD HH:MM:SS",
-        ),
-        "projection_y_coordinate": (
-            "feature_projection_y_coordinate",
-            "y position of the feature in the projection given by ProjectionCoordinateSystem",
-        ),
-        "projection_x_coordinate": (
-            "feature_projection_x_coordinate",
-            "x position of the feature in the projection given by ProjectionCoordinateSystem",
-        ),
-        "lat": ("feature_latitude", "latitude of the feature"),
-        "lon": ("feature_longitude", "longitude of the feature"),
-        "ncells": (
-            "feature_ncells",
-            "number of grid cells for this feature (meaning uncertain)",
-        ),
-        "areas": ("feature_area",),
-        "isolated": ("feature_isolation_flag",),
-        "num_objects": ("number_of_feature_neighbors",),
-        "cell": ("feature_parent_cell_id",),
-        "time_cell": ("feature_parent_cell_elapsed_time",),
-        "segmentation_mask": ("2d segmentation mask",),
-    }
-    new_feature_var_names = {
-        k: feature_standard_names[k][0]
-        for k in feature_standard_names.keys()
-        if k in TrackedFeatures.variables.keys()
-    }
-
-    TrackedFeatures = TrackedFeatures.drop(["cell_parent_track_id"])
-    # Combine Track and Mask variables. Use the 'feature' variable as the coordinate variable instead of
-    # the 'index' variable and call the dimension 'feature'
-    ds = xr.merge(
-        [
-            TrackedFeatures.swap_dims({"index": "feature"})
-            .drop("index")
-            .rename_vars(new_feature_var_names),
-            Mask,
-        ]
-    )
-
-    # Add the projection data back in
-    if not None in Projection:
-        ds["ProjectionCoordinateSystem"] = Projection
-
-    # Convert the cell ID variable from float to integer
-    if "int" not in str(TrackedFeatures.cell.dtype):
-        # The raw output from the tracking is actually an object array
-        # array([nan, 2, 3], dtype=object)
-        # (and is cast to a float array when saved as NetCDF, I think).
-        # Cast to float.
-        int_cell = xr.zeros_like(TrackedFeatures.cell, dtype="int64")
-
-        cell_id_data = TrackedFeatures.cell.astype("float64")
-        valid_cell = np.isfinite(cell_id_data)
-        valid_cell_ids = cell_id_data[valid_cell]
-        if not (np.unique(valid_cell_ids) > 0).all():
-            raise AssertionError(
-                "Lowest cell ID cell is less than one, conflicting with use of zero to indicate an untracked cell"
-            )
-        int_cell[valid_cell] = valid_cell_ids.astype("int64")
-        # ds['feature_parent_cell_id'] = int_cell
-    return ds
-
-
-def merge_split(TRACK, distance=25000, frame_len=5):
+def merge_split(TRACK, distance=25000, frame_len=5, dxy=500):
     """
     function to  postprocess tobac track data for merge/split cells
 
 
     Parameters
     ----------
-    TRACK : xarray.core.dataset.Dataset
-        xarray dataset of tobac Track information
+    TRACK : pandas.core.frame.DataFrame
+        Pandas dataframe of tobac Track information
 
     distance : float, optional
         Distance threshold prior to adding a pair of points into the minimum spanning tree.
@@ -235,6 +26,10 @@ def merge_split(TRACK, distance=25000, frame_len=5):
     frame_len : float, optional
         Threshold for the spanning length within which two points can be separated.
         Default is five (5) frames.
+
+    dxy : float, optional
+        The x/y/ grid spacing of the data.
+        Default is 500m.
 
     Returns
     -------
@@ -252,7 +47,38 @@ def merge_split(TRACK, distance=25000, frame_len=5):
 
     """
 
-    print("this is an update 3")
+    try:
+        import geopy
+    except ImportError:
+        geopy = None
+
+    if geopy:
+        from geopy.distance import geodesic
+    else:
+        print(
+            "Geopy not available, Merge/Split will proceed in tobac general coordinates."
+        )
+
+    try:
+        import networkx as nx
+    except ImportError:
+        networkx = None
+
+    #     if networkx:
+    #         import networkx as nx
+    #     else:
+    #         print("Cannot Merge/Split. Please install networkx.")
+    import logging
+    import numpy as np
+    from pandas.core.common import flatten
+    import xarray as xr
+
+    # Check if dxy is in meters of in kilometers. It should be in meters
+    if dxy <= 5:
+        dxy *= 1000
+
+    # Immediately convert pandas dataframe of track information to xarray:
+    TRACK = TRACK.to_xarray()
     track_groups = TRACK.groupby("cell")
     cell_ids = {cid: len(v) for cid, v in track_groups.groups.items()}
     id_data = np.fromiter(cell_ids.keys(), dtype=int)
@@ -283,11 +109,11 @@ def merge_split(TRACK, distance=25000, frame_len=5):
 
         for i in id_data:
             # print(i)
-            a_pointx = track_groups[i].projection_x_coordinate[-1].values
-            a_pointy = track_groups[i].projection_y_coordinate[-1].values
+            a_pointx = track_groups[i].hdim_2[-1].values * dxy
+            a_pointy = track_groups[i].hdim_1[-1].values * dxy
             for j in id_data:
-                b_pointx = track_groups[j].projection_x_coordinate[0].values
-                b_pointy = track_groups[j].projection_y_coordinate[0].values
+                b_pointx = track_groups[j].hdim_2[0].values * dxy
+                b_pointy = track_groups[j].hdim_1[0].values * dxy
                 d = np.linalg.norm(
                     np.array((a_pointx, a_pointy)) - np.array((b_pointx, b_pointy))
                 )
@@ -299,6 +125,8 @@ def merge_split(TRACK, distance=25000, frame_len=5):
                     b_names.append(j)
 
     id = []
+    # This is removing any tracks that have matched to themselves - e.g.,
+    # a beginning of a track has found its tail.
     for i in range(len(dist) - 1, -1, -1):
         if a_names[i] == b_names[i]:
             id.append(i)
@@ -310,12 +138,12 @@ def merge_split(TRACK, distance=25000, frame_len=5):
         else:
             continue
 
-    g = Graph()
+    g = nx.Graph()
     for i in np.arange(len(dist)):
         g.add_edge(a_names[i], b_names[i], weight=dist[i])
 
-    tree = minimum_spanning_edges(g)
-    tree_list = list(minimum_spanning_edges(g))
+    tree = nx.minimum_spanning_edges(g)
+    tree_list = list(tree)
 
     new_tree = []
     for i, j in enumerate(tree_list):
@@ -327,7 +155,7 @@ def merge_split(TRACK, distance=25000, frame_len=5):
 
     TRACK["cell_parent_track_id"] = np.zeros(len(TRACK["cell"].values))
     cell_id = np.unique(
-        TRACK.cell.values.astype(float)[~np.isnan(TRACK.cell.values.astype(float))]
+        TRACK.cell.values.astype(int)[~np.isnan(TRACK.cell.values.astype(int))]
     )
     track_id = dict()  # same size as number of total merged tracks
 
@@ -366,23 +194,15 @@ def merge_split(TRACK, distance=25000, frame_len=5):
 
                 track_id[np.nanmax(np.unique(temp))] = list(np.unique(temp))
 
-    storm_id = [0]  # default because we don't track larger storm systems *yet*
-    print("found storm id")
-
-    track_parent_storm_id = np.repeat(
-        0, len(track_id)
-    )  # This will always be zero when we don't track larger storm systems *yet*
-    print("found track parent storm ids")
-
     track_ids = np.array(list(track_id.keys()))
-    print("found track ids")
+    logging.debug("found track ids")
 
     cell_id = list(
         np.unique(
-            TRACK.cell.values.astype(float)[~np.isnan(TRACK.cell.values.astype(float))]
+            TRACK.cell.values.astype(int)[~np.isnan(TRACK.cell.values.astype(int))]
         )
     )
-    print("found cell ids")
+    logging.debug("found cell ids")
 
     cell_parent_track_id = []
 
@@ -395,18 +215,14 @@ def merge_split(TRACK, distance=25000, frame_len=5):
             cell_parent_track_id.append(np.repeat(int(id), len(track_id[int(id)])))
 
     cell_parent_track_id = list(flatten(cell_parent_track_id))
-    print("found cell parent track ids")
+    logging.debug("found cell parent track ids")
 
-    feature_parent_cell_id = list(TRACK.cell.values.astype(float))
-
-    print("found feature parent cell ids")
+    feature_parent_cell_id = list(TRACK.cell.values.astype(int))
+    logging.debug("found feature parent cell ids")
 
     # This version includes all the feature regardless of if they are used in cells or not.
     feature_id = list(TRACK.feature.values.astype(int))
-    print("found feature ids")
-
-    feature_parent_storm_id = np.repeat(0, len(feature_id))  # we don't do storms atm
-    print("found feature parent storm ids")
+    logging.debug("found feature ids")
 
     feature_parent_track_id = []
     feature_parent_track_id = np.zeros(len(feature_id))
@@ -420,58 +236,48 @@ def merge_split(TRACK, distance=25000, frame_len=5):
             trackid = cell_parent_track_id[j]
             feature_parent_track_id[i] = trackid
 
-    print("found feature parent track ids")
-
-    storm_child_track_count = [len(track_id)]
-    print("found storm child track count")
+    logging.debug("found feature parent track ids")
 
     track_child_cell_count = []
     for i, id in enumerate(track_id):
         track_child_cell_count.append(len(track_id[int(id)]))
-    print("found track child cell count")
+    logging.debug("found track child cell count")
 
     cell_child_feature_count = []
     for i, id in enumerate(cell_id):
         cell_child_feature_count.append(len(track_groups[id].feature.values))
-    print("found cell child feature count")
+    logging.debug("found cell child feature count")
 
-    storm_child_cell_count = [len(cell_id)]
-    storm_child_feature_count = [len(feature_id)]
-
-    storm_dim = "nstorms"
-    track_dim = "ntracks"
-    cell_dim = "ncells"
-    feature_dim = "nfeatures"
+    track_dim = "tracks"
+    cell_dim = "cells"
+    feature_dim = "features"
 
     d = xr.Dataset(
         {
-            "storm_id": (storm_dim, storm_id),
-            "track_id": (track_dim, track_ids),
-            "track_parent_storm_id": (track_dim, track_parent_storm_id),
-            "cell_id": (cell_dim, cell_id),
+            "track": (track_dim, track_ids),
+            "cell": (cell_dim, cell_id),
             "cell_parent_track_id": (cell_dim, cell_parent_track_id),
-            "feature_id": (feature_dim, feature_id),
+            "feature": (feature_dim, feature_id),
             "feature_parent_cell_id": (feature_dim, feature_parent_cell_id),
             "feature_parent_track_id": (feature_dim, feature_parent_track_id),
-            "feature_parent_storm_id": (feature_dim, feature_parent_storm_id),
-            "storm_child_track_count": (storm_dim, storm_child_track_count),
-            "storm_child_cell_count": (storm_dim, storm_child_cell_count),
-            "storm_child_feature_count": (storm_dim, storm_child_feature_count),
             "track_child_cell_count": (track_dim, track_child_cell_count),
             "cell_child_feature_count": (cell_dim, cell_child_feature_count),
         }
     )
-    d = d.set_coords(["feature_id", "cell_id", "track_id", "storm_id"])
 
-    assert len(track_id) == len(track_parent_storm_id)
+    d = d.set_coords(["feature", "cell", "track"])
+
     assert len(cell_id) == len(cell_parent_track_id)
     assert len(feature_id) == len(feature_parent_cell_id)
-    assert sum(storm_child_track_count) == len(track_id)
-    assert sum(storm_child_cell_count) == len(cell_id)
-    assert sum(storm_child_feature_count) == len(feature_id)
     assert sum(track_child_cell_count) == len(cell_id)
-    assert sum(
-        [sum(cell_child_feature_count), (len(np.where(feature_parent_track_id < 0)[0]))]
-    ) == len(feature_id)
+    assert (
+        sum(
+            [
+                sum(cell_child_feature_count[1:]),
+                (len(np.where(feature_parent_track_id < 0)[0])),
+            ]
+        )
+        == len(feature_id)
+    )
 
     return d
